% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetBuilder.R
\name{buildDataset}
\alias{buildDataset}
\title{Build CSV dataset from hierachical image collections}
\usage{
buildDataset(basedir, labels, isGray)
}
\arguments{
\item{basedir}{base directory}

\item{labels}{list of labels}

\item{isGray}{boolean flag whether convert image to gray or not}
}
\description{
Build CSV dataset from hierachical image collections
}

