% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapExtractor.R
\name{extractMap}
\alias{extractMap}
\title{Extract map patch from google maps}
\usage{
extractMap(basedir, df, truth, start, trialinaday)
}
\arguments{
\item{basedir}{base directory}

\item{df}{SpatialPointsDataFrame for target points}

\item{truth}{boolean selector positive negative position selection}

\item{start}{image index}

\item{trialinaday}{max trial in a day of google map query}
}
\description{
Extract map patch from google maps
}

