% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatasetBuilder.R
\name{imageToDataset}
\alias{imageToDataset}
\title{Convert images to CSV dataset}
\usage{
imageToDataset(basedir, outputdir, label, isGray)
}
\arguments{
\item{basedir}{base directory}

\item{outputdir}{directory for dataset file for each label}

\item{label}{label}

\item{isGray}{boolean flag whether convert image to gray or not}
}
\value{
converted data.frame of labeled images
}
\description{
Convert images to CSV dataset
}

