% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestMXNet.R
\name{TestMXNet}
\alias{TestMXNet}
\title{Test CNN model with MXNet}
\source{
TrainMXNet.R
}
\usage{
TestMXNet(basedir, isGray)
}
\arguments{
\item{basedir}{base directory}

\item{isGray}{isGray boolean flag whether convert image to gray or not}
}
\description{
Test CNN model with MXNet
}

