% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainMXNet.R
\name{trainCNN}
\alias{trainCNN}
\title{Train CNN model with MXNet}
\usage{
trainCNN(train.x, train.y, dim, channels)
}
\arguments{
\item{train.x}{training data.}

\item{train.y}{training labels}

\item{dim}{dimension of training data}

\item{channels}{the number of channels (color = 3, grayscale = 1)}
}
\value{
the model trained
}
\description{
\code{trainCNN} returns trained model with DNN
}

