% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainMXNet.R
\name{trainDNN}
\alias{trainDNN}
\title{Train DNN model with MXNet}
\usage{
trainDNN(train.x, train.y)
}
\arguments{
\item{train.x}{training data}

\item{train.y}{training labels}
}
\value{
the model trained
}
\description{
\code{trainDNN} returns trained model with DNN
}

